backend/requirements.txt
# backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0

